feat: implement schema querying for oracledb
