+++ apps/api/package.json
@@ -49,7 +49,6 @@
     "lru-cache": "^10.2.2",
     "mysql2": "^3.10.1",
     "node-fetch": "^3.3.2",
-    "oracledb": "^6.5.1",
     "p-all": "^5.0.0",
     "p-queue": "^8.0.1",
     "parse-duration": "^1.1.0",
@@ -85,7 +84,6 @@
     "@types/multer": "^1.4.10",
     "@types/node": "^20.9.0",
     "@types/node-fetch": "^2.6.9",
-    "@types/oracledb": "^6.5.0",
     "@types/ramda": "^0.29.9",
     "@types/semver": "^7.5.8",
     "@types/split2": "^4.2.3",
